This is effectively a complete rewrite of the globaldce algorithm, resulting in it being both shorter and more effective.  It no longer depends on the callgraph, so one FIXME has been fixed.
Additionally, this pass was not able to delete recursive (but dead) functions
if they were pointed to by global variables which were also dead.  In fact
this pass had a lot of problems deleting functions which were only pointed
to by dead globals and other stuff.

Fixing this means that the entire EH library should be stripped away now from
programs that don't use sjlj or exceptions.


git-svn-id: https://llvm.org/svn/llvm-project@8567 91177308-0d34-0410-b5e6-96231b3b80d8