Cleanup and restructure the code to make it easier to read and maintain. The only functionality change is that we now implement:   Regression/Transforms/TailCallElim/intervening-inst.ll
Which is really kinda pointless, because it means that trivially dead code
does not interfere with -tce, but trivially dead code probably wouldn't be
around anytime when this pass is run anyway.

The point of including this change it to support other more aggressive
transformations when we have the analysis capabilities to do so.


git-svn-id: https://llvm.org/svn/llvm-project@10312 91177308-0d34-0410-b5e6-96231b3b80d8