After unrolling our single basic block loop, fold it into the preheader and exit block.  The primary motivation for doing this is that we can now unroll nested loops.
This makes a pretty big difference in some cases.  For example, in 183.equake,
we are now beating the native compiler with the CBE, and we are a lot closer
with LLC.

I'm now going to play around a bit with the unroll factor and see what effect
it really has.


git-svn-id: https://llvm.org/svn/llvm-project@13034 91177308-0d34-0410-b5e6-96231b3b80d8