Fix an annoying Mac bug: notebook tab changes don't show up on the display.
Move TVTreeCtrl's EVENT_TABLE up, nearer to the rest of the TVTreeCtrl
implementation.

Move the Explanation down, nearer to where it is used.

Fix a bug where if you selected a DSGraph then a Code view or Text view, the
text view would be constrained to the size of the DSGraph.


git-svn-id: https://llvm.org/svn/llvm-project@13539 91177308-0d34-0410-b5e6-96231b3b80d8