Rewrite the branch selector to be correct in the face of large functions. The algorithm it used before wasn't 100% correct, we now use an iterative expansion model.  This fixes assembler errors when compiling 403.gcc with tail merging enabled.
Change the way the branch selector works overall: Now, the isel generates
PPC::BCC instructions (as it used to) directly, and these BCC instructions
are emitted to the output or jitted directly if branches don't need
expansion.  Only if branches need expansion are instructions rewritten
and created.  This should make branch select faster, and eliminates the
Bxx instructions from the .td file.


git-svn-id: https://llvm.org/svn/llvm-project@31837 91177308-0d34-0410-b5e6-96231b3b80d8