Bug #: Submitted by: Reviewed by: Fixed an "impurity" in the typechecking code (to more closely model the C99 spec)...
- Added function Sema::DefaultFunctionArrayConversion(). This is called
by UsualUnaryConversions(). It is also called from several contexts that
don't call for integer promotions (logical negation for expressions, and
while/if/do/for conditions in statements). It is also used in
UsualAssignmentConversions (which is cleaner than the check I was using
before).

- Changed the name from UsualUnaryConversion->UsualUnaryConversions.

- Added comments to the header.

Since several contexts don't call for integer
promotions,


git-svn-id: https://llvm.org/svn/llvm-project@39581 91177308-0d34-0410-b5e6-96231b3b80d8