Forward port of [129262].
A function that only has cleanups (destructors) to run when an exception is raised
currently is not assigned a personality function because we don't bother creating
a call to eh.selector if there are no catches.  But the personality function is
still needed, since it is the personality function that analyses the exception 
table
and decides where to jump to (to the cleanups in this case).  This patch causes a
call to eh.selector to be generated in every landing pad, even if it has no type 
infos,
because that specifies the personality function.

Patch from Duncan Sands.


git-svn-id: https://llvm.org/svn/llvm-project@39808 91177308-0d34-0410-b5e6-96231b3b80d8