Add code to make test/CodeGen/struct-init.c work correctly without the pointer cast hack currently in isIntegerConstantExpr (in lib/AST/Expr.cpp).  Also removes an odd test that even gcc doesn't accept.
The reason the pointer cast hack is relevant here is that it makes Sema 
end up misinterpreting the relevant expression as a null pointer constant.

The reason for this patch is that I plan to remove the pointer cast hack 
sometime soon because it causes strange issues, especially in its 
current form; see my recent email to cfe-dev
"[PATCH] add constant expression evaluation to the AST and fix PR2413".



git-svn-id: https://llvm.org/svn/llvm-project@52120 91177308-0d34-0410-b5e6-96231b3b80d8