Pulling r53361 into Gaz:
Move the IsVolatile and SVOffset fields into the MemSDNode base
class, and store IsVolatile and Alignment in a more compact form.
This makes AtomicSDNode slightly larger, but it shrinks LoadSDNode
and StoreSDNode, which are much more common and are the largest of
the SDNode subclasses. Also, this lets the isVolatile() and
getAlignment() accessors be non-virtual.



git-svn-id: https://llvm.org/svn/llvm-project@53366 91177308-0d34-0410-b5e6-96231b3b80d8