This is an early implementation of a vertical waymarking algorithm. The idea is to eliminate Use::Val, thus shrinking struct Use to 8 bytes.
Of course this needs some recalculation logic
to recover the Value* by walking the def/use
chain. Whether the computing overhead outweighs
the space savings is an open question at this
point.

There are some tests at the end, so far all
pass.

Please Note: this is a private branch, this file
would break the llvm build, I'll fix this later.


git-svn-id: https://llvm.org/svn/llvm-project@56160 91177308-0d34-0410-b5e6-96231b3b80d8